requirements.txt
discord.py==2.3.2
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0 